refactor describecollectionprovider
